test(elysia): split tests by function and add basic coverage
