+++ packages/elysia/package.json
@@ -9,8 +9,9 @@
     "build": "pnpm sync:modules && tsdown",
     "lint": "oxlint",
     "lint:fix": "oxlint --fix",
-    "test": "vitest --run --project stateless",
-    "test:all": "vitest --run",
+    "test": "vitest --run",
+    "test:stateful": "vitest --run --project stateful",
+    "test:stateless": "vitest --run --project stateless",
     "test:watch": "vitest",
     "test:coverage": "vitest --run --coverage",
     "format": "oxfmt",
